+++ src/app/admin/create/page.tsx
@@ -1,317 +1,40 @@
 "use client"
 
-import type React from "react"
-import { useState, useEffect } from "react"
-import { Building2, Plus, X, Calendar, MapPin, Briefcase, Upload, FileUp, FileText, Loader2 } from "lucide-react"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Textarea } from "@/components/ui/textarea"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Checkbox } from "@/components/ui/checkbox"
-import { DatePicker } from "@/components/ui/date-picker"
+import { useState } from "react"
 import { createJobPosting } from "@/api/posting"
-import { getPositions, createPosition } from "@/api/posting"
 import { toast } from "sonner"
 import { useRouter } from "next/navigation"
 import { 
   AttachmentFile, 
-  CompanyLocation, 
   CreateJobPostingRequest, 
-  EmploymentType, 
-  Position 
 } from "@/types/posting"
-import usePostingFileUpload from "@/hooks/usePostingFileUpload"
-import { formatFileSize } from "@/utils/formatFileSize"
-
-const ALLOWED_EXTENSIONS = [
-  'pdf', 'jpeg', 'jpg', 'png', 'xls', 'xlsx', 'xlsm', 'hwp', 'hwpx', 'hwt', 'ppt', 'pptx', 'zip'
-] as const;
-
-type AllowedExtension = typeof ALLOWED_EXTENSIONS[number];
-
-const MAX_FILE_SIZE = 10 * 1024 * 1024;
-
-const LOCATION_MAPPING: Record<string, CompanyLocation> = {
-  "서울": "SEOUL",
-  "부산": "BUSAN",
-  "대구": "DAEGU",
-  "인천": "INCHEON",
-  "광주": "GWANGJU",
-  "대전": "DAEJEON",
-  "울산": "ULSAN",
-  "세종": "SEJONG",
-  "경기": "GYEONGGI",
-  "강원": "GANGWON",
-  "충북": "CHUNGBUK",
-  "충남": "CHUNGNAM",
-  "전북": "JEONBUK",
-  "전남": "JEONNAM",
-  "경북": "GYEONGBUK",
-  "경남": "GYEONGNAM",
-  "제주": "JEJU",
-};
-
-const AVAILABLE_LOCATIONS = Object.keys(LOCATION_MAPPING);
-
-const EMPLOYMENT_TYPE_MAPPING: Record<string, EmploymentType> = {
-  "fulltime": "FULL_TIME",
-  "contract": "CONTRACT",
-  "intern": "INTERN",
-  "military": "MILITARY_EXCEPTION"
-};
-
-interface JobFormData {
-  company: string;
-  description: string;
-  location: string;
-  customLocation: string;
-  employmentType: string;
-  startDate: Date | null;
-  endDate: Date | null;
-  attachments: AttachmentWithFile[];
-}
-
-interface AttachmentWithFile {
-  file: File;
-  url?: string;
-}
-
-interface FileValidationResult {
-  isValid: boolean;
-  errorMessage?: string;
-}
+import JobForm, { JobFormData, EMPLOYMENT_TYPE_MAPPING, LOCATION_MAPPING } from "@/components/JobForm"
 
 export default function CreateJobPage() {
   const router = useRouter();
-  const { uploadFile, isUploading } = usePostingFileUpload();
-  
-  const [positions, setPositions] = useState<Position[]>([]);
-  const [selectedPositions, setSelectedPositions] = useState<number[]>([]);
-  const [customPositionName, setCustomPositionName] = useState("");
-  const [isPositionsLoading, setIsPositionsLoading] = useState(true);
-  const [isAddingPosition, setIsAddingPosition] = useState(false);
-  const [isDragging, setIsDragging] = useState(false);
   const [isSubmitting, setIsSubmitting] = useState(false);
-  
-  const [jobData, setJobData] = useState<JobFormData>({
-    company: "",
-    description: "",
-    location: "",
-    customLocation: "",
-    employmentType: "",
-    startDate: null,
-    endDate: null,
-    attachments: [],
-  });
 
-  useEffect(() => {
-    const loadPositions = async () => {
-      try {
-        const positionsData = await getPositions();
-        setPositions(positionsData);
-      } catch (error) {
-        console.error(error);
-        toast.error("포지션 데이터를 불러오는데 실패했습니다.");
-      } finally {
-        setIsPositionsLoading(false);
-      }
-    };
-
-    loadPositions();
-  }, []);
-
-  const handleInputChange = <K extends keyof JobFormData>(field: K, value: JobFormData[K]) => {
-    setJobData(prev => ({ ...prev, [field]: value }));
-  };
-
-  const handleDateChange = (field: "startDate" | "endDate", date: Date | undefined) => {
-    handleInputChange(field, date === undefined ? null : date);
-  };
-
-  const togglePosition = (positionId: number, checked: boolean) => {
-    if (checked) {
-      setSelectedPositions(prev => [...prev, positionId]);
-    } else {
-      setSelectedPositions(prev => prev.filter(id => id !== positionId));
-    }
-  };
-
-  const addCustomPosition = async () => {
-    if (!customPositionName.trim()) return;
-    
-    try {
-      setIsAddingPosition(true);
-      const newPosition = await createPosition(customPositionName);
-      
-      setPositions(prev => [...prev, newPosition]);
-      setSelectedPositions(prev => [...prev, newPosition.id]);
-      setCustomPositionName("");
-      
-      toast.success(`'${newPosition.name}' 포지션이 추가되었습니다.`);
-    } catch (error) {
-      console.error(error);
-      toast.error("포지션 추가에 실패했습니다.");
-    } finally {
-      setIsAddingPosition(false);
-    }
-  };
-
-  const validateFile = (file: File): FileValidationResult => {
-    const fileExtension = file.name.split('.').pop()?.toLowerCase() as AllowedExtension | undefined;
-    
-    if (!fileExtension || !ALLOWED_EXTENSIONS.includes(fileExtension as AllowedExtension)) {
-      return { 
-        isValid: false, 
-        errorMessage: `'${file.name}'은(는) 지원하지 않는 확장자입니다.` 
-      };
-    }
-
-    if (file.size > MAX_FILE_SIZE) {
-      return { 
-        isValid: false, 
-        errorMessage: `'${file.name}'의 크기가 10MB를 초과합니다.` 
-      };
-    }
-
-    return { isValid: true };
-  };
-
-  const handleFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(e.target.files || []);
-    if (files.length === 0) return;
-
-    for (const file of files) {
-      const validationResult = validateFile(file);
-      if (!validationResult.isValid) {
-        toast.error(validationResult.errorMessage);
-        return;
-      }
-    }
-
-    const newAttachments = [...jobData.attachments];
-    
-    for (const file of files) {
-      const isDuplicate = newAttachments.some(attachment => 
-        attachment.file.name === file.name && attachment.file.size === file.size
-      );
-      
-      if (!isDuplicate) {
-        newAttachments.push({ file });
-      }
-    }
-    
-    handleInputChange("attachments", newAttachments);
-  };
-
-  const removeFile = (index: number) => {
-    const newFiles = jobData.attachments.filter((_, i) => i !== index);
-    handleInputChange("attachments", newFiles);
-  };
-
-  const handleDragEnter = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const files = e.dataTransfer.files;
-    if (files && files.length > 0) {
-      const event = {
-        target: {
-          files: files
-        }
-      } as unknown as React.ChangeEvent<HTMLInputElement>;
-      handleFileChange(event);
-    }
-  };
-
-  const handleFileUploadClick = () => {
-    document.getElementById('file-upload')?.click();
-  };
-
-  const validateRequiredFields = (): boolean => {
-    return Boolean(
-      jobData.company &&
-      jobData.description &&
-      selectedPositions.length > 0 &&
-      jobData.location &&
-      jobData.employmentType &&
-      jobData.startDate &&
-      jobData.endDate
-    );
-  };
-
-  const processFileUploads = async (): Promise<AttachmentFile[]> => {
-    if (jobData.attachments.length === 0) return [];
-    
-    try {
-      const uploadPromises = jobData.attachments.map(async (attachment) => {
-        if (attachment.url) {
-          return {
-            url: attachment.url,
-            name: attachment.file.name
-          };
-        }
-        
-        const url = await uploadFile(attachment.file);
-        return {
-          url,
-          name: attachment.file.name
-        };
-      });
-
-      const uploadedFiles = await Promise.all(uploadPromises);
-      toast.success('모든 파일이 성공적으로 업로드되었습니다.');
-      return uploadedFiles;
-    } catch (error) {
-      console.error("파일 업로드 실패:", error);
-      toast.error("파일 업로드에 실패했습니다");
-      throw error;
-    }
-  };
-
-  const handleSubmit = async () => {  
+  const handleSubmit = async (
+    formData: JobFormData, 
+    selectedPositions: number[], 
+    uploadedFiles: AttachmentFile[]
+  ) => {
     try {
       setIsSubmitting(true);
       
-      if (!validateRequiredFields()) {
-        toast.error("모든 필수 항목을 입력해주세요");
-        return;
-      }
-      
-      const finalLocation = jobData.location === "custom" ? jobData.customLocation : jobData.location;
-      const mappedLocation = jobData.location === "custom" 
+      const finalLocation = formData.location === "custom" ? formData.customLocation : formData.location;
+      const mappedLocation = formData.location === "custom" 
         ? "SEOUL"
         : LOCATION_MAPPING[finalLocation] || "SEOUL";
       
-      const uploadedFiles = await processFileUploads();
-      
-      const startAt = jobData.startDate ? jobData.startDate.toISOString() : "";
-      const endAt = jobData.endDate ? jobData.endDate.toISOString() : "";
+      const startAt = formData.startDate ? formData.startDate.toISOString() : "";
+      const endAt = formData.endDate ? formData.endDate.toISOString() : "";
       
       const requestData: CreateJobPostingRequest = {
-        companyName: jobData.company,
-        companyDescription: jobData.description,
+        companyName: formData.company,
+        companyDescription: formData.description,
         companyLocation: mappedLocation,
-        employmentType: EMPLOYMENT_TYPE_MAPPING[jobData.employmentType],
+        employmentType: EMPLOYMENT_TYPE_MAPPING[formData.employmentType],
         positions: selectedPositions,
         files: uploadedFiles,
         startAt,
@@ -331,285 +54,6 @@ export default function CreateJobPage() {
     }
   };
 
-  const renderBasicInfoSection = () => (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2">
-          <Building2 className="h-5 w-5" />
-          기본 정보
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        <div>
-          <Label htmlFor="company">회사명 *</Label>
-          <Input
-            id="company"
-            value={jobData.company}
-            onChange={(e) => handleInputChange("company", e.target.value)}
-            placeholder="회사명을 입력하세요"
-          />
-        </div>
-
-        <div>
-          <Label htmlFor="description">회사 소개 및 채용 내용 *</Label>
-          <Textarea
-            id="description"
-            value={jobData.description}
-            onChange={(e) => handleInputChange("description", e.target.value)}
-            placeholder="회사 소개, 업무 내용, 자격 요건, 우대 사항 등을 상세히 작성해주세요"
-            rows={8}
-          />
-        </div>
-      </CardContent>
-    </Card>
-  );
-
-  const renderPositionSection = () => (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2">
-          <Briefcase className="h-5 w-5" />
-          모집 포지션
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        {isPositionsLoading ? (
-          <div className="py-4 flex justify-center">
-            <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-600"></div>
-          </div>
-        ) : (
-          <>
-            <div>
-              <Label>포지션 선택</Label>
-              <div className="grid grid-cols-2 md:grid-cols-3 gap-2 mt-2">
-                {positions.map((position) => (
-                  <div key={position.id} className="flex items-center space-x-2">
-                    <Checkbox
-                      id={`position-${position.id}`}
-                      checked={selectedPositions.includes(position.id)}
-                      onCheckedChange={(checked) => togglePosition(position.id, !!checked)}
-                    />
-                    <Label htmlFor={`position-${position.id}`} className="text-sm">
-                      {position.name}
-                    </Label>
-                  </div>
-                ))}
-              </div>
-            </div>
-
-            <div>
-              <Label htmlFor="customPosition">커스텀 포지션 추가</Label>
-              <div className="flex gap-2 mt-2">
-                <Input
-                  id="customPosition"
-                  value={customPositionName}
-                  onChange={(e) => setCustomPositionName(e.target.value)}
-                  placeholder="새로운 포지션명 입력"
-                  disabled={isAddingPosition}
-                />
-                <Button 
-                  onClick={addCustomPosition} 
-                  disabled={!customPositionName || isAddingPosition}
-                >
-                  {isAddingPosition ? (
-                    <div className="animate-spin h-4 w-4 border-2 border-white rounded-full border-t-transparent" />
-                  ) : (
-                    <Plus className="h-4 w-4" />
-                  )}
-                </Button>
-              </div>
-            </div>
-          </>
-        )}
-      </CardContent>
-    </Card>
-  );
-
-  const renderWorkConditionsSection = () => (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2">
-          <MapPin className="h-5 w-5" />
-          근무 조건
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-          <div>
-            <Label htmlFor="location">근무 지역 *</Label>
-            <Select value={jobData.location} onValueChange={(value) => handleInputChange("location", value)}>
-              <SelectTrigger>
-                <SelectValue placeholder="지역을 선택하세요" />
-              </SelectTrigger>
-              <SelectContent>
-                {AVAILABLE_LOCATIONS.map((location) => (
-                  <SelectItem key={location} value={location}>
-                    {location}
-                  </SelectItem>
-                ))}
-                <SelectItem value="custom">직접 입력</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-
-          <div>
-            <Label htmlFor="employmentType">고용 형태 *</Label>
-            <Select
-              value={jobData.employmentType}
-              onValueChange={(value) => handleInputChange("employmentType", value)}
-            >
-              <SelectTrigger>
-                <SelectValue placeholder="고용 형태를 선택하세요" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="fulltime">정규직</SelectItem>
-                <SelectItem value="contract">계약직</SelectItem>
-                <SelectItem value="intern">인턴</SelectItem>
-                <SelectItem value="military">병역특례</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        </div>
-
-        {jobData.location === "custom" && (
-          <div>
-            <Label htmlFor="customLocation">직접 입력</Label>
-            <Input
-              id="customLocation"
-              value={jobData.customLocation}
-              onChange={(e) => handleInputChange("customLocation", e.target.value)}
-              placeholder="근무 지역을 입력하세요"
-            />
-          </div>
-        )}
-      </CardContent>
-    </Card>
-  );
-
-  const renderApplicationPeriodSection = () => (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2">
-          <Calendar className="h-5 w-5" />
-          지원 기간
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-          <div>
-            <Label>시작일 *</Label>
-            <DatePicker 
-              date={jobData.startDate} 
-              onDateChange={(date) => handleDateChange("startDate", date)} 
-            />
-          </div>
-
-          <div>
-            <Label>마감일 *</Label>
-            <DatePicker 
-              date={jobData.endDate} 
-              onDateChange={(date) => handleDateChange("endDate", date)} 
-            />
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-
-  const renderAttachmentsSection = () => (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2">
-          <FileText className="h-5 w-5" />
-          첨부 파일
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        <div>
-          <Label htmlFor="attachments">파일 업로드</Label>
-          <div
-            className={`mt-2 flex justify-center px-6 pt-5 pb-6 border-2 ${
-              isDragging ? 'border-blue-400 bg-blue-50' : 'border-gray-300'
-            } border-dashed rounded-lg hover:border-blue-400 transition-colors cursor-pointer`}
-            onDragEnter={handleDragEnter}
-            onDragOver={handleDragOver}
-            onDragLeave={handleDragLeave}
-            onDrop={handleDrop}
-            onClick={handleFileUploadClick}
-          >
-            <div className="space-y-1 text-center">
-              <Upload className="mx-auto h-8 w-8 text-gray-400" />
-              <div className="flex text-sm text-gray-600 items-center justify-center">
-                <span className="flex items-center gap-1">
-                  <FileUp className="h-4 w-4" />
-                  파일 선택
-                </span>
-                <input
-                  id="file-upload"
-                  name="file-upload"
-                  type="file"
-                  multiple
-                  className="sr-only"
-                  onChange={handleFileChange}
-                />
-                <p className="pl-1">또는 드래그 앤 드롭</p>
-              </div>
-              <p className="text-xs text-gray-500">
-                {ALLOWED_EXTENSIONS.join(', ')} 파일만 업로드 가능 (최대 10MB)
-              </p>
-            </div>
-          </div>
-        </div>
-
-        {jobData.attachments.length > 0 && (
-          <div className="space-y-3 mt-4">
-            <Label>업로드된 파일</Label>
-            {jobData.attachments.map((attachment, index) => (
-              <Card key={index}>
-                <CardContent className="p-4">
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center gap-3">
-                      {isUploading ? (
-                        <Loader2 className="h-5 w-5 text-gray-400 animate-spin" />
-                      ) : (
-                        <FileText className="h-5 w-5 text-gray-400" />
-                      )}
-                      <div>
-                        <p className="font-medium">{attachment.file.name}</p>
-                        <p className="text-sm text-gray-500">
-                          {formatFileSize(attachment.file.size)}
-                        </p>
-                      </div>
-                    </div>
-                    <Button variant="ghost" size="sm" onClick={() => removeFile(index)}>
-                      <X className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </CardContent>
-              </Card>
-            ))}
-          </div>
-        )}
-      </CardContent>
-    </Card>
-  );
-
-  const renderActionButtons = () => (
-    <div className="flex justify-end gap-4">
-      <Button 
-        onClick={handleSubmit}
-        disabled={
-          isSubmitting ||
-          isUploading ||
-          !validateRequiredFields()
-        }
-        className="text-white bg-blue-600 hover:bg-blue-700"
-      >
-        {isSubmitting ? "처리 중..." : "공고 등록"}
-      </Button>
-    </div>
-  );
-
   return (
     <div className="min-h-screen bg-gray-50">
       <main className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
@@ -618,14 +62,11 @@ export default function CreateJobPage() {
           <p className="text-gray-600">새로운 채용공고를 등록하여 GSM 학생들에게 기회를 제공하세요</p>
         </div>
 
-        <div className="grid gap-6">
-          {renderBasicInfoSection()}
-          {renderPositionSection()}
-          {renderWorkConditionsSection()}
-          {renderApplicationPeriodSection()}
-          {renderAttachmentsSection()}
-          {renderActionButtons()}
-        </div>
+        <JobForm
+          submitButtonText="공고 등록"
+          onSubmit={handleSubmit}
+          isSubmitting={isSubmitting}
+        />
       </main>
     </div>
   )
